Partial link & include debug symbols in .o files
clang on Mac OS X is very helpful and when run in a single pass to both
compile and link, it automatically runs dsymutil which splits out debug
symbols into a .dSYM directory. We don't want that. In particular
because we are just building a .o file and not a final executable.

It appears the only way around it is to split the cc invokation into two
separate commands; one for compiling .c and another for partial linking.

Partial linking also means that dependencies on other libraries are
resolved at compile time. Thus we no longer have run time dependencies
on these libraries.