Generate GC descriptors
We now generate a GC descriptor for structs. It describes the memory
structure, where there are pointers, to the GC. The descriptor itself is
stored in the _class struct in the first field. The old $GCINFO  was
meant for this type of GC information but we never got around to using
it before, instead it held a string of the name of the object class -
useful for debugging. We still think that's useful so we keep the name
but it is now stored in $name - slightly more intuitive.

The GC descriptor is built from a bitmap. As it can vary in size, it is
kept outside of the _class struct as a new global of the name ClassD_gcbm.